+++ demos/newapp/newapp/apps/blog/models/models.py
@@ -2,6 +2,7 @@ from sqlalchemy import (
     Column,
     Integer,
     Text,
+    Unicode
     )
 
 from . import DBSession, Base
@@ -9,7 +10,7 @@ from . import DBSession, Base
 class Post(Base):
     __tablename__ = 'posts'
     id = Column(Integer, primary_key=True)
-    title = Column(Text, unique=True)
+    title = Column(Unicode, unique=True)
     content = Column(Text)
 
     def __init__(self, title, content):