+++ demos/newapp/newapp/models/models.py
@@ -2,6 +2,7 @@ from sqlalchemy import (
     Column,
     Integer,
     Text,
+    Unicode
     )
 
 from . import DBSession, Base
@@ -9,7 +10,7 @@ from . import DBSession, Base
 class Site(Base):
     __tablename__ = 'site'
     id = Column(Integer, primary_key=True)
-    name = Column(Text, unique=True)
+    name = Column(Unicode, unique=True)
     value = Column(Integer)
 
     def __init__(self, name, value):